Extract Results to its own component, use tabs to navigate between indexes on smaller screens
