+++ src/components/Search.tsx
@@ -3,29 +3,11 @@
 import { InformationCircleIcon } from '@heroicons/react/24/outline';
 import Image from 'next/image';
 import { useEffect, useState } from 'react';
-import { Index, useSearchBox } from 'react-instantsearch-hooks-web';
-import HitList from './HitList';
+import { useSearchBox } from 'react-instantsearch-hooks-web';
+import Results from './Results';
 import SearchBox from './SearchBox';
 
 const queryParam = 'q';
-
-const indexes = [
-  {
-    name: 'docs',
-    title: 'Docs',
-    link: 'https://docs.convex.dev',
-  },
-  {
-    name: 'stack',
-    title: 'Stack',
-    link: 'https://stack.convex.dev',
-  },
-  {
-    name: 'discord',
-    title: 'Discord',
-    link: 'https://discord.com/invite/nk6C2qTeCq',
-  },
-];
 
 export default function Search() {
   const { refine } = useSearchBox();
@@ -69,7 +51,7 @@ export default function Search() {
   }, []);
 
   return (
-    <div className="flex min-h-screen flex-col">
+    <div className="flex min-h-screen flex-col gap-6">
       <header className="sticky top-0 z-10 flex h-32 flex-col justify-center gap-4 border-b border-neutral-n10 bg-neutral-n12 px-4 md:h-20 md:flex-row md:items-center md:gap-12">
         <Image src="/logo.svg" alt="Convex logo" width={320} height={36} />
         <SearchBox
@@ -78,35 +60,19 @@ export default function Search() {
           onClear={handleClear}
         />
       </header>
-      {query === '' ? (
-        <div className="flex grow items-center justify-center gap-2 p-4">
-          <InformationCircleIcon className="w-8 shrink-0 text-green-g4" />
-          <span className="text-neutral-n4 md:text-xl">
-            Use the input above to search across Docs, Stack, and Discord.
-          </span>
-        </div>
-      ) : (
-        <main className="flex grow flex-col gap-12 p-4 xl:flex-row xl:gap-6">
-          {indexes.map(({ name, title, link }) => (
-            <div
-              key={name}
-              className="flex grow basis-0 flex-col md:max-w-lg xl:overflow-hidden"
-            >
-              <a
-                href={link}
-                className="sticky top-32 mt-4 bg-neutral-n12 py-4 font-display text-2xl leading-none text-neutral-n2 underline-offset-4 shadow-lg stretch-max hover:underline md:top-2 xl:static xl:mt-0 xl:text-xl xl:font-bold"
-                target="_blank"
-              >
-                {title}
-              </a>
-              <Index indexName={name}>
-                <HitList />
-              </Index>
-            </div>
-          ))}
-        </main>
-      )}
-      <footer className="mt-6 flex flex-col-reverse justify-between gap-4 border-t border-neutral-n10 p-6 sm:flex-row">
+      <main className="flex grow px-4">
+        {query === '' ? (
+          <div className="flex w-full items-center justify-center gap-2">
+            <InformationCircleIcon className="w-8 shrink-0 text-green-g4" />
+            <span className="text-neutral-n4 md:text-xl">
+              Use the input above to search across Docs, Stack, and Discord.
+            </span>
+          </div>
+        ) : (
+          <Results />
+        )}
+      </main>
+      <footer className="flex flex-col-reverse justify-between gap-4 border-t border-neutral-n10 px-4 py-6 sm:flex-row">
         <span className="text-neutral-n4">Copyright © 2023 Convex, Inc.</span>
         <div className="flex gap-4">
           <a
